refactor: rename 'update-pending-invoices.ts'
