+++ core/api/src/app/wallets/index.ts
@@ -12,10 +12,10 @@ export * from "./get-transaction-by-journal-id"
 export * from "./get-transactions-by-addresses"
 export * from "./get-transactions-by-hash"
 export * from "./get-transactions-for-wallet"
+export * from "./handle-held-invoices"
 export * from "./register-broadcasted-payout-txn"
 export * from "./settle-payout-txn"
 export * from "./update-legacy-on-chain-receipt"
-export * from "./update-pending-invoices"
 export * from "./validate"
 export * from "./get-invoice-for-wallet-by-hash"
 export * from "./get-pending-incoming-on-chain-transactions-for-wallets"
